Refactor: Reestructuracion de submodulos 'change-password' e 'inicio'
