+++ src/app/(modulos)/(autenticacion)/inicio/(presentation)/components/Inicio.tsx
@@ -1,4 +1,4 @@
-"use client";
+﻿"use client";
 
 import React, { useEffect } from 'react';
 import { useRouter } from 'next/navigation';
